test: move tests into example folder and add granular interactivity tests
