+++ example/.gitignore
@@ -23,3 +23,6 @@ dist-ssr
 *.sln
 *.sw?
 *.ignore
+
+# Test output
+tests/results
