+++ example/playwright.config.mts
@@ -1,6 +1,5 @@
 /* eslint no-console: 0 */
-import { devices } from '@playwright/test';
-import type { PlaywrightTestConfig } from '@playwright/test';
+import { type PlaywrightTestConfig, devices } from '@playwright/test';
 
 const IS_CI = !!process.env.CI;
 const PORT = !isNaN(Number(process.env.PORT)) ? Number(process.env.PORT) : 3000;
@@ -12,16 +11,15 @@ console.info('[playwright.config]', { IS_CI, PORT });
 const config: PlaywrightTestConfig = {
   forbidOnly: IS_CI,
   retries: IS_CI ? 2 : 0,
-  testDir: './test',
-  outputDir: './test/results',
+  testDir: './',
+  outputDir: './tests/results',
   reporter: 'list',
   webServer: {
-    command: 'pnpm run build && pnpm --filter "react-image-turntable-example" run start',
+    command: 'pnpm run start',
     reuseExistingServer: !IS_CI,
     url: `http://localhost:${PORT}`,
     env: {
       NODE_ENV: 'test',
-      USE_BABEL_PLUGIN_ISTANBUL: '1',
     },
   },
   use: {
@@ -35,15 +33,15 @@ const config: PlaywrightTestConfig = {
       name: 'chromium',
       use: { ...devices['Desktop Chrome'] },
     },
-    // {
-    //   name: 'firefox',
-    //   use: { ...devices['Desktop Firefox'] },
-    // },
-    // {
-    //   name: 'webkit',
-    //   use: { ...devices['Desktop Safari'] },
-    // },
-  ],
+    IS_CI && {
+      name: 'firefox',
+      use: { ...devices['Desktop Firefox'] },
+    },
+    IS_CI && {
+      name: 'webkit',
+      use: { ...devices['Desktop Safari'] },
+    },
+  ].filter(Boolean) as PlaywrightTestConfig['projects'],
 };
 
 export default config;
