+++ lib/package.json
@@ -1,5 +1,5 @@
 {
-  "version": "3.1.0",
+  "version": "3.1.2",
   "name": "react-image-turntable",
   "license": "MIT",
   "repository": {
@@ -71,19 +71,20 @@
     "react-dom": ">=16.8"
   },
   "devDependencies": {
-    "@types/node": "^20.8.9",
-    "@types/react": "^18.2.33",
-    "@types/react-dom": "^18.2.14",
-    "browserslist": "^4.22.1",
-    "concurrently": "^8.2.2",
+    "@arethetypeswrong/cli": "^0.16.2",
+    "@types/node": "^22.5.5",
+    "@types/react": "^18.3.5",
+    "@types/react-dom": "^18.3.0",
+    "browserslist": "^4.23.3",
+    "concurrently": "^9.0.1",
     "esbuild-plugin-babel": "^0.2.3",
-    "esbuild-plugin-browserslist": "^0.10.0",
-    "publint": "^0.2.5",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "release-it": "^17.0.0",
-    "tslib": "~2.6.2",
-    "tsup": "^8.0.1",
-    "typescript": "^5.2.2"
+    "esbuild-plugin-browserslist": "^0.14.0",
+    "publint": "^0.2.10",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "release-it": "^17.6.0",
+    "tslib": "~2.7.0",
+    "tsup": "^8.2.4",
+    "typescript": "^5.6.2"
   }
 }
