+++ pnpm-workspace.yaml
@@ -1,3 +1,3 @@
 packages:
-  - 'lib/**'
   - 'example/**'
+  - 'lib/**'
